requirements.txt
certifi==2020.6.20
coloredlogs==15.0.1
cycler==0.10.0
decorator==4.4.2
flatbuffers==23.1.4
humanfriendly==10.0
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.2.0
matplotlib==3.3.2
mpmath==1.2.1
networkx==2.5
numpy==1.21.6
onnx==1.13.0
onnxruntime==1.13.1
opencv-python==4.4.0.44
packaging==23.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.20.3
pydantic==1.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.2
SimpleITK==2.0.0
six==1.15.0
sklearn==0.0
sympy==1.10.1
threadpoolctl==2.1.0
tifffile==2020.9.22
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.7.4.3